chore: Refactor OpenApiEditor and SwaggerEditor to reduce duplicate code
