+++ samtranslator/swagger/swagger.py
@@ -1,14 +1,15 @@
 ﻿import copy
 import re
-from typing import Dict, Any, Iterator
+from typing import Dict, Any, Optional
 
 from samtranslator.model.apigateway import ApiGatewayAuthorizer
-from samtranslator.model.intrinsics import ref, make_conditional, fnSub, is_intrinsic_no_value
+from samtranslator.model.intrinsics import ref, make_conditional, fnSub
 from samtranslator.model.exceptions import InvalidDocumentException, InvalidTemplateException
+from samtranslator.open_api.base_editor import BaseEditor
 from samtranslator.utils.py27hash_fix import Py27Dict, Py27UniStr
 
 
-class SwaggerEditor(object):
+class SwaggerEditor(BaseEditor):
     """
     Wrapper class capable of parsing and generating Swagger JSON.  This implements Swagger spec just enough that SAM
     cares about. It is built to handle "partial Swagger" ie. Swagger that is incomplete and won't
@@ -23,25 +24,20 @@ class SwaggerEditor(object):
     """
 
     _OPTIONS_METHOD = "options"
-    _X_APIGW_INTEGRATION = "x-amazon-apigateway-integration"
     _X_APIGW_BINARY_MEDIA_TYPES = "x-amazon-apigateway-binary-media-types"
-    _CONDITIONAL_IF = "Fn::If"
     _X_APIGW_GATEWAY_RESPONSES = "x-amazon-apigateway-gateway-responses"
     _X_APIGW_POLICY = "x-amazon-apigateway-policy"
-    _X_ANY_METHOD = "x-amazon-apigateway-any-method"
     _X_APIGW_REQUEST_VALIDATORS = "x-amazon-apigateway-request-validators"
     _X_APIGW_REQUEST_VALIDATOR = "x-amazon-apigateway-request-validator"
     _X_ENDPOINT_CONFIG = "x-amazon-apigateway-endpoint-configuration"
     _CACHE_KEY_PARAMETERS = "cacheKeyParameters"
     # https://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html
-    _ALL_HTTP_METHODS = ["OPTIONS", "GET", "HEAD", "POST", "PUT", "DELETE", "PATCH"]
     _EXCLUDED_PATHS_FIELDS = ["summary", "description", "parameters"]
     _POLICY_TYPE_IAM = "Iam"
     _POLICY_TYPE_IP = "Ip"
     _POLICY_TYPE_VPC = "Vpc"
-    _SERVERS = "servers"
 
-    def __init__(self, doc):  # type: ignore[no-untyped-def]
+    def __init__(self, doc: Optional[Dict[str, Any]]) -> None:
         """
         Initialize the class with a swagger dictionary. This class creates a copy of the Swagger and performs all
         modifications on this copy.
@@ -50,7 +46,7 @@ class SwaggerEditor(object):
         :raises InvalidDocumentException: If the input Swagger document does not meet the basic Swagger requirements.
         """
 
-        if not SwaggerEditor.is_valid(doc):  # type: ignore[no-untyped-call]
+        if not doc or not SwaggerEditor.is_valid(doc):
             raise InvalidDocumentException([InvalidTemplateException("Invalid Swagger document")])
 
         self._doc = copy.deepcopy(doc)
@@ -66,67 +62,8 @@ class SwaggerEditor(object):
         # We can do an early path validation on path item objects,
         # so we don't need to validate wherever we use them.
         for path in self.iter_on_path():
-            for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
+            for path_item in self.get_conditional_contents(self.paths.get(path)):
                 SwaggerEditor.validate_path_item_is_dict(path_item, path)
-
-    def get_conditional_contents(self, item):  # type: ignore[no-untyped-def]
-        """
-        Returns the contents of the given item.
-        If a conditional block has been used inside the item, returns a list of the content
-        inside the conditional (both the then and the else cases). Skips {'Ref': 'AWS::NoValue'} content.
-        If there's no conditional block, then returns an list with the single item in it.
-
-        :param dict item: item from which the contents will be extracted
-        :return: list of item content
-        """
-        contents = [item]
-        if isinstance(item, dict) and self._CONDITIONAL_IF in item:
-            contents = item[self._CONDITIONAL_IF][1:]
-            contents = [content for content in contents if not is_intrinsic_no_value(content)]
-        return contents
-
-    def has_path(self, path, method=None):  # type: ignore[no-untyped-def]
-        """
-        Returns True if this Swagger has the given path and optional method
-        For paths with conditionals, only returns true if both items (true case, and false case) have the method.
-
-        :param string path: Path name
-        :param string method: HTTP method
-        :return: True, if this path/method is present in the document
-        """
-        if path not in self.paths:
-            return False
-
-        method = self._normalize_method_name(method)  # type: ignore[no-untyped-call]
-        if method:
-            for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
-                if method not in path_item:
-                    return False
-        return True
-
-    def method_has_integration(self, method):  # type: ignore[no-untyped-def]
-        """
-        Returns true if the given method contains a valid method definition.
-        This uses the get_conditional_contents function to handle conditionals.
-
-        :param dict method: method dictionary
-        :return: true if method has one or multiple integrations
-        """
-        for method_definition in self.get_conditional_contents(method):  # type: ignore[no-untyped-call]
-            if self.method_definition_has_integration(method_definition):  # type: ignore[no-untyped-call]
-                return True
-        return False
-
-    def method_definition_has_integration(self, method_definition):  # type: ignore[no-untyped-def]
-        """
-        Checks a method definition to make sure it has an apigw integration
-
-        :param dict method_definition: method definition dictionary
-        :return: True if an integration exists
-        """
-        if method_definition.get(self._X_APIGW_INTEGRATION):
-            return True
-        return False
 
     @staticmethod
     def _update_dict(obj: Dict[Any, Any], k: str, v: Dict[Any, Any]) -> None:
@@ -155,41 +92,6 @@ class SwaggerEditor(object):
         else:
             SwaggerEditor._update_dict(self._doc, self._X_ENDPOINT_CONFIG, set_disable_api_endpoint)
 
-    def has_integration(self, path, method):  # type: ignore[no-untyped-def]
-        """
-        Checks if an API Gateway integration is already present at the given path/method.
-        For paths with conditionals, it only returns True if both items (true case, false case) have the integration
-
-        :param string path: Path name
-        :param string method: HTTP method
-        :return: True, if an API Gateway integration is already present
-        """
-        method = self._normalize_method_name(method)  # type: ignore[no-untyped-call]
-
-        if not self.has_path(path, method):  # type: ignore[no-untyped-call]
-            return False
-
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
-            method_definition = path_item.get(method)
-            if not (isinstance(method_definition, dict) and self.method_has_integration(method_definition)):  # type: ignore[no-untyped-call]
-                return False
-        # Integration present and non-empty
-        return True
-
-    def add_path(self, path, method=None):  # type: ignore[no-untyped-def]
-        """
-        Adds the path/method combination to the Swagger, if not already present
-
-        :param string path: Path name
-        :param string method: HTTP method
-        """
-        method = self._normalize_method_name(method)  # type: ignore[no-untyped-call]
-
-        self.paths.setdefault(path, Py27Dict())
-
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
-            path_item.setdefault(method, Py27Dict())
-
     def add_lambda_integration(  # type: ignore[no-untyped-def]
         self, path, method, integration_uri, method_auth_config=None, api_auth_config=None, condition=None
     ):
@@ -201,8 +103,8 @@ class SwaggerEditor(object):
         :param string integration_uri: URI for the integration.
         """
 
-        method = self._normalize_method_name(method)  # type: ignore[no-untyped-call]
-        if self.has_integration(path, method):  # type: ignore[no-untyped-call]
+        method = self._normalize_method_name(method)
+        if self.has_integration(path, method):
             raise InvalidDocumentException(
                 [
                     InvalidTemplateException(
@@ -211,14 +113,14 @@ class SwaggerEditor(object):
                 ]
             )
 
-        self.add_path(path, method)  # type: ignore[no-untyped-call]
+        self.add_path(path, method)
 
         # Wrap the integration_uri in a Condition if one exists on that function
         # This is necessary so CFN doesn't try to resolve the integration reference.
         if condition:
             integration_uri = make_conditional(condition, integration_uri)
 
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
+        for path_item in self.get_conditional_contents(self.paths.get(path)):
             path_item[method][self._X_APIGW_INTEGRATION] = Py27Dict()
             # insert key one by one to preserce input order
             path_item[method][self._X_APIGW_INTEGRATION]["type"] = "aws_proxy"
@@ -271,20 +173,20 @@ class SwaggerEditor(object):
         :param bool condition: Condition for the integration
         """
 
-        method = self._normalize_method_name(method)  # type: ignore[no-untyped-call]
-        if self.has_integration(path, method):  # type: ignore[no-untyped-call]
+        method = self._normalize_method_name(method)
+        if self.has_integration(path, method):
             raise InvalidDocumentException(
                 [InvalidTemplateException("Integration already exists on Path={}, Method={}".format(path, method))]
             )
 
-        self.add_path(path, method)  # type: ignore[no-untyped-call]
+        self.add_path(path, method)
 
         # Wrap the integration_uri in a Condition if one exists on that state machine
         # This is necessary so CFN doesn't try to resolve the integration reference.
         if condition:
             integration_uri = make_conditional(condition, integration_uri)
 
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
+        for path_item in self.get_conditional_contents(self.paths.get(path)):
             # Responses
             integration_responses = Py27Dict()
             # insert key one by one to preserce input order
@@ -314,49 +216,13 @@ class SwaggerEditor(object):
             if condition:
                 path_item[method] = make_conditional(condition, path_item[method])
 
-    def make_path_conditional(self, path, condition):  # type: ignore[no-untyped-def]
-        """
-        Wrap entire API path definition in a CloudFormation if condition.
-        """
-        self.paths[path] = make_conditional(condition, self.paths[path])
-
     def _generate_integration_credentials(self, method_invoke_role=None, api_invoke_role=None):  # type: ignore[no-untyped-def]
         return self._get_invoke_role(method_invoke_role or api_invoke_role)  # type: ignore[no-untyped-call]
 
-    def _get_invoke_role(self, invoke_role):  # type: ignore[no-untyped-def]
+    @staticmethod
+    def _get_invoke_role(invoke_role):  # type: ignore[no-untyped-def]
         CALLER_CREDENTIALS_ARN = "arn:aws:iam::*:user/*"
         return invoke_role if invoke_role and invoke_role != "CALLER_CREDENTIALS" else CALLER_CREDENTIALS_ARN
-
-    def iter_on_path(self) -> Iterator[str]:
-        """
-        Yields all the paths available in the Swagger. As a caller, if you add new paths to Swagger while iterating,
-        they will not show up in this iterator
-
-        :yields string: Path name
-        """
-
-        for path, _ in self.paths.items():
-            yield path
-
-    def iter_on_method_definitions_for_path_at_method(  # type: ignore[no-untyped-def]
-        self, path_name, method_name, skip_methods_without_apigw_integration=True
-    ):
-        """
-        Yields all the method definitions for the path+method combinations if path and/or method have IF conditionals.
-        If there are no conditionals, will just yield the single method definition at the given path and method name.
-
-        :param path_name: path name
-        :param method_name: method name
-        :param skip_methods_without_apigw_integration: if True, skips method definitions without apigw integration
-        :yields dict: method definition
-        """
-        normalized_method_name = self._normalize_method_name(method_name)  # type: ignore[no-untyped-call]
-
-        for path_item in self.get_conditional_contents(self.paths.get(path_name)):  # type: ignore[no-untyped-call]
-            for method_definition in self.get_conditional_contents(path_item.get(normalized_method_name)):  # type: ignore[no-untyped-call]
-                if skip_methods_without_apigw_integration and not self.method_definition_has_integration(method_definition):  # type: ignore[no-untyped-call]
-                    continue
-                yield method_definition
 
     def iter_on_all_methods_for_path(self, path_name, skip_methods_without_apigw_integration=True):  # type: ignore[no-untyped-def]
         """
@@ -366,22 +232,24 @@ class SwaggerEditor(object):
         :param skip_methods_without_apigw_integration: if True, skips method definitions without apigw integration
         :yields list of (method name, method definition) tuples
         """
-        for path_item in self.get_conditional_contents(self.paths.get(path_name)):  # type: ignore[no-untyped-call]
+        for path_item in self.get_conditional_contents(self.paths.get(path_name)):
             for method_name, method in path_item.items():
                 # Excluding non-method sections
                 if method_name in SwaggerEditor._EXCLUDED_PATHS_FIELDS:
                     continue
 
-                for method_definition in self.get_conditional_contents(method):  # type: ignore[no-untyped-call]
+                for method_definition in self.get_conditional_contents(method):
                     SwaggerEditor.validate_is_dict(
                         method_definition,
                         'Value of "{}" ({}) for path {} is not a valid dictionary.'.format(
                             method_name, method_definition, path_name
                         ),
                     )
-                    if skip_methods_without_apigw_integration and not self.method_definition_has_integration(method_definition):  # type: ignore[no-untyped-call]
+                    if skip_methods_without_apigw_integration and not self.method_definition_has_integration(
+                        method_definition
+                    ):
                         continue
-                    normalized_method_name = self._normalize_method_name(method_name)  # type: ignore[no-untyped-call]
+                    normalized_method_name = self._normalize_method_name(method_name)
                     yield normalized_method_name, method_definition
 
     def add_cors(  # type: ignore[no-untyped-def]
@@ -411,9 +279,9 @@ class SwaggerEditor(object):
         :raises InvalidTemplateException: When values for one of the allowed_* variables is empty
         """
 
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
+        for path_item in self.get_conditional_contents(self.paths.get(path)):
             # Skip if Options is already present
-            method = self._normalize_method_name(self._OPTIONS_METHOD)  # type: ignore[no-untyped-call]
+            method = self._normalize_method_name(self._OPTIONS_METHOD)
             if method in path_item:
                 continue
 
@@ -588,7 +456,7 @@ class SwaggerEditor(object):
         for authorizer_name, authorizer in authorizers.items():
             self.security_definitions[authorizer_name] = authorizer.generate_swagger()
 
-    def add_awsiam_security_definition(self):  # type: ignore[no-untyped-def]
+    def add_awsiam_security_definition(self) -> None:
         """
         Adds AWS_IAM definition to the securityDefinitions part of Swagger.
         Note: this method is idempotent
@@ -609,7 +477,7 @@ class SwaggerEditor(object):
         if "AWS_IAM" not in self.security_definitions:
             self.security_definitions.update(aws_iam_security_definition)
 
-    def add_apikey_security_definition(self):  # type: ignore[no-untyped-def]
+    def add_apikey_security_definition(self) -> None:
         """
         Adds api_key definition to the securityDefinitions part of Swagger.
         Note: this method is idempotent
@@ -722,9 +590,9 @@ class SwaggerEditor(object):
                 # The first element of the method_definition['security'] should be AWS_IAM
                 # because authorizer_list = ['AWS_IAM'] is hardcoded above
                 if "AWS_IAM" in method_definition["security"][0]:
-                    self.add_awsiam_security_definition()  # type: ignore[no-untyped-call]
+                    self.add_awsiam_security_definition()
 
-    def set_path_default_apikey_required(self, path):  # type: ignore[no-untyped-def]
+    def set_path_default_apikey_required(self, path: str) -> None:
         """
         Add the ApiKey security as required for each method on this path unless ApiKeyRequired
         was defined at the Function/Path/Method level. This is intended to be used to set the
@@ -823,7 +691,7 @@ class SwaggerEditor(object):
         if authorizers is None:
             authorizers = Py27Dict()
 
-        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):  # type: ignore[no-untyped-call]
+        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):
 
             security_dict = Py27Dict()
             security_dict[authorizer_name] = []
@@ -855,7 +723,7 @@ class SwaggerEditor(object):
                 # The first element of the method_definition['security'] should be AWS_IAM
                 # because authorizer_list = ['AWS_IAM'] is hardcoded above
                 if "AWS_IAM" in method_definition["security"][0]:
-                    self.add_awsiam_security_definition()  # type: ignore[no-untyped-call]
+                    self.add_awsiam_security_definition()
 
     def _set_method_apikey_handling(self, path, method_name, apikey_required):  # type: ignore[no-untyped-def]
         """
@@ -866,14 +734,14 @@ class SwaggerEditor(object):
         :param string method_name: Method name
         :param bool apikey_required: Whether the apikey security is required
         """
-        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):  # type: ignore[no-untyped-call]
+        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):
 
             if apikey_required:
                 # We want to enable apikey required security
                 security_dict = Py27Dict()
                 security_dict["api_key"] = []
                 apikey_security = [security_dict]
-                self.add_apikey_security_definition()  # type: ignore[no-untyped-call]
+                self.add_apikey_security_definition()
             else:
                 # The method explicitly does NOT require apikey and there is an API default
                 # so let's add a marker 'api_key_false' so that we don't incorrectly override
@@ -919,7 +787,7 @@ class SwaggerEditor(object):
             # Adding only if the validator hasn't been defined already
             self._doc[self._X_APIGW_REQUEST_VALIDATORS].update(request_validator_definition)
 
-        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):  # type: ignore[no-untyped-call]
+        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):
             set_validator_to_method = Py27Dict({self._X_APIGW_REQUEST_VALIDATOR: validator_name})
             # Setting validator to the given method
             method_definition.update(set_validator_to_method)
@@ -935,7 +803,7 @@ class SwaggerEditor(object):
         model_name = request_model and request_model.get("Model").lower()
         model_required = request_model and request_model.get("Required")
 
-        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):  # type: ignore[no-untyped-call]
+        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):
             if self._doc.get("swagger") is not None:
 
                 existing_parameters = method_definition.get("parameters", [])
@@ -955,7 +823,7 @@ class SwaggerEditor(object):
                 method_definition["parameters"] = existing_parameters
 
             elif self._doc.get("openapi") and SwaggerEditor.safe_compare_regex_with_string(
-                SwaggerEditor.get_openapi_version_3_regex(), self._doc["openapi"]
+                SwaggerEditor._OPENAPI_VERSION_3_REGEX, self._doc["openapi"]
             ):
                 method_definition["requestBody"] = {
                     "content": {"application/json": {"schema": {"$ref": "#/components/schemas/{}".format(model_name)}}}
@@ -1124,7 +992,7 @@ class SwaggerEditor(object):
         The regex removes the trailing slash to ensure the permission works as intended
         """
         methods = []
-        for path_item in self.get_conditional_contents(self.paths.get(path)):  # type: ignore[no-untyped-call]
+        for path_item in self.get_conditional_contents(self.paths.get(path)):
             methods += list(path_item.keys())
 
         uri_list = []
@@ -1279,7 +1147,7 @@ class SwaggerEditor(object):
         :return:
         """
 
-        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):  # type: ignore[no-untyped-call]
+        for method_definition in self.iter_on_method_definitions_for_path_at_method(path, method_name):
             existing_parameters = method_definition.get("parameters", [])
 
             for request_parameter in request_parameters:
@@ -1311,7 +1179,7 @@ class SwaggerEditor(object):
             method_definition["parameters"] = existing_parameters
 
     @property
-    def swagger(self):  # type: ignore[no-untyped-def]
+    def swagger(self) -> Dict[str, Any]:
         """
         Returns a **copy** of the Swagger document as a dictionary.
 
@@ -1334,7 +1202,7 @@ class SwaggerEditor(object):
         return copy.deepcopy(self._doc)
 
     @staticmethod
-    def is_valid(data):  # type: ignore[no-untyped-def]
+    def is_valid(data: Any) -> bool:
         """
         Checks if the input data is a Swagger document
 
@@ -1347,38 +1215,13 @@ class SwaggerEditor(object):
                 return True
             if bool(data.get("openapi")):
                 return SwaggerEditor.safe_compare_regex_with_string(
-                    SwaggerEditor.get_openapi_version_3_regex(), data["openapi"]
+                    SwaggerEditor._OPENAPI_VERSION_3_REGEX, data["openapi"]
                 )
         return False
 
     @staticmethod
     def validate_open_api_version_3(api_version: str) -> bool:
-        return SwaggerEditor.safe_compare_regex_with_string(SwaggerEditor.get_openapi_version_3_regex(), api_version)
-
-    @staticmethod
-    def validate_is_dict(obj: Any, exception_message: str) -> None:
-        """
-        Throws exception if obj is not a dict
-
-        :param obj: object being validated
-        :param exception_message: message to include in exception if obj is not a dict
-        """
-
-        if not isinstance(obj, dict):
-            raise InvalidDocumentException([InvalidTemplateException(exception_message)])
-
-    @staticmethod
-    def validate_path_item_is_dict(path_item: Any, path: str) -> None:
-        """
-        Throws exception if path_item is not a dict
-
-        :param path_item: path_item (value at the path) being validated
-        :param path: path name
-        """
-
-        SwaggerEditor.validate_is_dict(
-            path_item, "Value of '{}' path must be a dictionary according to Swagger spec.".format(path)
-        )
+        return SwaggerEditor.safe_compare_regex_with_string(SwaggerEditor._OPENAPI_VERSION_3_REGEX, api_version)
 
     @staticmethod
     def gen_skeleton() -> Py27Dict:
@@ -1396,49 +1239,9 @@ class SwaggerEditor(object):
         return skeleton
 
     @staticmethod
-    def _get_authorization_scopes(authorizers: Dict[str, ApiGatewayAuthorizer], default_authorizer: str) -> Any:
-        """
-        Returns auth scopes for an authorizer if present
-        :param authorizers: authorizer definitions
-        :param default_authorizer: name of the default authorizer
-        """
-        authorizer = authorizers.get(default_authorizer)
-        if authorizer and authorizer.authorization_scopes is not None:
-            return authorizer.authorization_scopes
-        return []
-
-    @staticmethod
-    def _normalize_method_name(method):  # type: ignore[no-untyped-def]
-        """
-        Returns a lower case, normalized version of HTTP Method. It also know how to handle API Gateway specific methods
-        like "ANY"
-
-        NOTE: Always normalize before using the `method` value passed in as input
-
-        :param string method: Name of the HTTP Method
-        :return string: Normalized method name
-        """
-        if not method or not isinstance(method, str):
-            return method
-
-        method = method.lower()
-        if method == "any":
-            return SwaggerEditor._X_ANY_METHOD
-        return method
-
-    @staticmethod
     def get_openapi_versions_supported_regex() -> str:
         openapi_version_supported_regex = r"\A[2-3](\.\d)(\.\d)?$"
         return openapi_version_supported_regex
-
-    @staticmethod
-    def get_openapi_version_3_regex() -> str:
-        openapi_version_3_regex = r"\A3(\.\d)(\.\d)?$"
-        return openapi_version_3_regex
-
-    @staticmethod
-    def safe_compare_regex_with_string(regex: str, data: str) -> bool:
-        return re.match(regex, str(data)) is not None
 
     @staticmethod
     def get_path_without_trailing_slash(path):  # type: ignore[no-untyped-def]
